requirements.txt
# Core processing
numpy>=1.24.0
pandas>=2.0.0
pydub>=0.25.1
librosa>=0.10.0
scipy>=1.10.0

# Audio transcription
openai-whisper>=20231117
whisperx>=3.0.0  # For improved diarization

# Translation and NLP
openai>=1.3.0
transformers>=4.35.0

# TTS
elevenlabs>=0.2.24

# Web interface
streamlit>=1.28.0

# Development tools
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0

# Environment variables
python-dotenv>=0.19.0